rag-with-postgresql/pyproject.toml
[project]
name = "rag-with-postgresql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.12.0",
    "google-genai>=1.25.0",
    "python-dotenv>=1.1.1",
    "kagglehub>=0.3.13",
    "openpyxl>=3.1.5",
    "pgvector>=0.4.1",
    "plotly>=6.3.0",
    "pyarrow>=21.0.0",
    "pydantic-settings>=2.10.1",
    "sentence-transformers>=5.1.0",
    "sqlalchemy[asyncio]>=2.0.42",
    "transformers",
    "typing-inspect>=0.9.0",
    "umap-learn>=0.5.9.post2",
    "pandas>=2.3.2",
    "asyncpg>=0.30.0",
]
